requirements.txt
black==22.1.0
click==8.0.4
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
tk==0.1.0
tomli==2.0.1
typing_extensions==4.1.1